Add files via upload
Used automobile information is included in the dataset. The information can be utilized for a variety of tasks, including price prediction, which serves as an example of how machine learning uses linear regression. Preprocessing the data has been done to get it into the proper shape so that the model may be trained. To get the highest accuracy, a method is developed following preprocessing and model training.